Add bldidx and uses utilities. bldidx reads each INC file, looking for -PUBLIC declarations. These are extracted into a composite JSON document with functions, variables, operators, classes and methods.
The uses utility reads standard input. If a variable or function
matches one in the IDX file created by bldidx, an -INCLUDE is
generated. Note that only functions and variables are processed
(not classes, methods, or operators).

Each lss file updated to add -PUBLIC lines. JIT.lss is not yet
completed.